pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["arseniy <snow_2015@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.2"
aiogram = "^3.6.0"
openai = "^1.30.5"
pydantic-settings = "^2.2.1"
openai-whisper = {git = "https://github.com/openai/whisper"}
async-sqlalchemy = "^1.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
amplitude-analytics = "^1.1.4"

[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
